Fix mismatched types between addr and offset in mmap_pmemfile_t.rs

+++ capybaraKV/capybarakv/src/pmem/mmap_pmemfile_t.rs
@@ -191,7 +191,7 @@ impl PersistentMemoryRegion for FileBackedPersistentMemoryRegion {
         MaybeCorruptedBytes<S>,
         PmemError,
     >) where S: PmCopy {
-        let addr = addr + self.section.offset;
+        let addr = addr + self.section.offset as u64;
         let mut mmf_borrowed = self.section.mmf.borrow_mut();
         let pm_slice: &[u8] = &mmf_borrowed.mmap[addr as usize..addr as usize + S::size_of()];
 
@@ -212,7 +212,7 @@ impl PersistentMemoryRegion for FileBackedPersistentMemoryRegion {
 
     #[verifier::external_body]
     fn read_unaligned(&self, addr: u64, num_bytes: u64) -> (bytes: Result<Vec<u8>, PmemError>) {
-        let addr = addr + self.section.offset;
+        let addr = addr + self.section.offset as u64;
         let mut mmf_borrowed = self.section.mmf.borrow_mut();
         let pm_slice: &[u8] = &mmf_borrowed.mmap[addr as usize..(addr + num_bytes) as usize];
 
@@ -224,7 +224,7 @@ impl PersistentMemoryRegion for FileBackedPersistentMemoryRegion {
 
     #[verifier::external_body]
     fn write(&mut self, addr: u64, bytes: &[u8]) {
-        let addr = addr + self.section.offset;
+        let addr = addr + self.section.offset as u64;
         let mut mmf_borrowed = self.section.mmf.borrow_mut();
         mmf_borrowed.mmap[addr as usize..addr as usize + bytes.len()].copy_from_slice(bytes);
     }
@@ -232,7 +232,7 @@ impl PersistentMemoryRegion for FileBackedPersistentMemoryRegion {
     #[verifier::external_body]
     #[allow(unused_variables)]
     fn serialize_and_write<S>(&mut self, addr: u64, to_write: &S) where S: PmCopy + Sized {
-        let addr = addr + self.section.offset;
+        let addr = addr + self.section.offset as u64;
         let num_bytes: usize = S::size_of() as usize;
 
         // convert the given &S to a pointer, then a slice of bytes
